Add action success/error notifications
Extract components into files and refactor
